Create a unique name for mongoDB collection at start of fuzzer
This commit allows users to run multiple instances of ODfuzz at the same time. For each instance of the fuzzer, a unique name of database collection is created. Because of this, data will not be overwritten from concurrently running fuzzers, and each instance of the fuzzer will write data to the different location.