Fix deleter
Pointer's data were overwritten after the part was deleted but the query string remained untouched. This led to unconsistent state where the query string was correctly built, but the data were equal to the last query that was generated. Such a behaviour happened only while generating multiple queries at once using the -a option.

Total hours wasted to solve this bug: 54