Do not generate orderby, skip, top query options for single entities
The query options like orderby, skip, and top should not be generated in queries which trigger GET_ENTITY methods. Also, it does not make any sense to run skip or top query on single entity. In this commit, there is also added a full support for GET_ENTITY method. In the past, GET_ENTITY methods were triggered only in non-addressable entities.