Fix indentation, add workaround for generating accessible path
When a query is built, a new accessible path is generated each time. Accessible path differentiates between requesting a single entity (e.g. CorrespondenceTypeSet(Event='SAP06',VariantId='SAP06',Id='SAP06')/) or multiple entities (e.g. CorrespondenceTypeSet/). A query that is mutated may contain additional skip or top query options which are not allowed in requests that access single entities. In this commit, we introduce a simple check to forbid generation of key values for accessible path.