Add unique HASH value for each URL
This commit enhances the way of processing the fuzzer's results. For each URL, there is generated a unique HASH value that represents an identifier for the corresponding URL. The hash is stored in the pivot table as well as in the list of all generated urls. Now, a user can search for the identifier in the file urls_list.txt, based on previous assumtions obtained from the pivot table, and examine a particular URL with less effort.