Check if a path of a certificate was specified
A status of a certificate was not determined correctly. Before
this commit, the property method has_certificate() used to
return True even if the certificate was not declared. The
exception TypeError (stat: path should be string, bytes,
os.PathLike or integer, not NoneType) was raised while the
fuzzer was trying to check if a path to the certificate
is a valid file.

https://github.com/SAP/odfuzz/blob/184660b47b2e17838b39922cc17e40aa2304a985/odfuzz/fuzzer.py#L1170-L1172