Encode string values right after the generation process
As of now, all strings, which may contain special characters, are automatically encoded. Due to that, while saving these strings to a database, it is necessary to decode them back again to prevent mutations for the encoded strings what may lead to discrepancies. This behavior can be disabled using the environment variable ODFUZZ_USE_ENCODER=False.

In this commit, there was also the updated class interface for Edm generators to improve overall readability.