Sync property values in URI and body for PUT
The fuzzed property values in payloads for PUT used to generate the values separately for the URI and the body, due to how generators worked.

Now, during key generation, generate() is called with a new format value (format == 'key') which returns a tuple (uri_value, body_value) of the property fuzzed in two formats, having the same value. The body_value is used for body generation.