Tests: make test_generators.py idempotent
This is ugly hack to already ugly code (too many globla variables and static/class methods). By adding reset() method to EncoderMixin, there is ability to set the default/init state of encoders in test teardown fixture, so other tests using the RandomGenerator would not be affected by persistant state, set up by last test in the module by pushing some function to RandomGenerator._encode property.

What did not work, even tho apparent:

Save the current functionin the property (even with deepcopy) to temp variable and set back in the test end. By some collision of pytest magic and @classmethod decorator, the tests would fail, because the "previous" encoder, stored in the temp var, is used in the generated string; even (not sure why) if the set back is after the assert line.
-

Goal to the future:
Refactor entirely in the core library, that would have better testability by as less global state as possible.