Fixing DataType Issues Giving 400 Bad Request
Made Changes to the three different data types:

1. Edm.Int64: At the end of the Int64 Digit there was a Character 'L' appended which was giving the 400 errors, so Just removed that concatenation with the digit and no more 400 errors where given.

2. Edm.Double: There were two changes to this Data Type.

A. Invalid Character: A character 'd' was appended to the end of the digit which was giving the error.

B. URL Encoding: The Character '+'/'-' in the range (2.23e-40, 1.19e+40) for decimal was URL encoded when sent to the endpoint which was giving the 400 Bad request.

3. Edm.Decimal: The error was regarding the scale given for the Decimal Digit, When the scale was 0 the Decimal number had no decimal point due to which the endpoint was giving a Data Type Error, so converted the Scale to a minimum of 2 as Scale of 2 is what accepted for a Decimal Digit.