Fixing DataType Issues Giving 400 Bad Request
Made Changes to the two different data types:

1. Edm.Int64: At the end of the Int64 Digit there was a Character 'L' appended which was giving the 400 errors, so Just removed that concatenation from the body and created a separate structure for URI and Body.

2. Edm.Double: There were two changes to this Data Type.

A. Invalid Character: A character 'd' was appended to the end of the digit which was giving the error.

B. URL Encoding: The Character '+'/'-' in the range (2.23e-40, 1.19e+40) for decimal was URL encoded when sent to the endpoint which was giving the 400 Bad request.

Created a separate data type structure for URI and Body for the request been sent