Fixing 400 bad request for Edm.Decimal
While Fuzzing any endpoint, there were many 400 Bad Requests given as response from the endpoint.
One of the Major reason being Type conversion errors for Edm.Decimal.

Edm.Decimal: The error was regarding the scale given for the Decimal Digit, this scale is been parsed from the metadata file for the endpoint which we have to fuzz and so scale decides how many digits will be there after the decimal point. When the scale was 0 the Decimal number had no digits after the decimal point due to which the endpoint was not considering the digit as a Decimal number due to which the endpoint was giving a Data Type Error.

Example:
Earliar: 82734682374682734682736482374
Now: 82734682374682734682736482374.00

Subsequently added a testcase for this particular scenario where the scale value should be 0 in test_generators.py file